DicomTagExplorer/requirements.txt
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.24.0
tkinterdnd2>=0.3.0
pylibjpeg[all]>=2.0.0
imageio[ffmpeg]>=2.30.0
